pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drift-analyzer"
version = "0.1.0"
description = "Detect architectural erosion from AI-generated code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = [
    "code-analysis",
    "technical-debt",
    "architecture",
    "ai-code",
    "static-analysis",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "gitpython>=3.1",
    "networkx>=3.0",
]

[project.optional-dependencies]
typescript = [
    "tree-sitter>=0.22",
    "tree-sitter-typescript>=0.22",
]
embeddings = [
    "sentence-transformers>=3.0",
    "faiss-cpu>=1.7",
    "numpy>=1.26",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]
all = [
    "drift-analyzer[typescript,embeddings,dev]",
]

[project.scripts]
drift = "drift.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/drift"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
